refactor(mcp): extract cartog-mcp test module into focused files
The ~2500-line #[cfg(test)] mod tests (111 tests) moves out of lib.rs into
src/tests/{update,schema,validation,handlers,single_writer,lsp_gate,
read_tools}.rs. lib.rs drops to 2085 lines. Tests stay in-crate (private
access via crate::*); snap and test_provider hoisted to the tests hub.
Per-test #[cfg(feature = "lsp")] preserved. No test added, dropped, or
modified.
