+++ crates/cartog-mcp/src/lib.rs
@@ -2082,2531 +2082,4 @@ pub(crate) use single_writer::{
 };
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    // ── parse_arm_output tests (cartog_update envelope reshaping) ──
-    //
-    // Unix-only: these construct a real `ExitStatus` by spawning `true`/`sh`.
-    // The parse logic itself is platform-independent.
-
-    /// Build an `Output` with the given stdout/stderr and a zero exit status.
-    /// parse_arm_output reads `status` only in the no-output branch, so a
-    /// success-shaped status is fine for the parse-path cases.
-    #[cfg(unix)]
-    fn output_ok(stdout: &str, stderr: &str) -> std::process::Output {
-        std::process::Output {
-            status: std::process::Command::new("true")
-                .status()
-                .expect("run true"),
-            stdout: stdout.as_bytes().to_vec(),
-            stderr: stderr.as_bytes().to_vec(),
-        }
-    }
-
-    #[test]
-    fn mcp_compact_defaults_on_and_parses_opt_out() {
-        let _g = test_validate_call_counter::SERIAL.blocking_lock();
-        let prev = std::env::var_os("CARTOG_MCP_COMPACT");
-
-        std::env::remove_var("CARTOG_MCP_COMPACT");
-        assert!(mcp_compact(), "compact is the default when unset");
-
-        for off in ["0", "false", "no", "off", "OFF", " false "] {
-            std::env::set_var("CARTOG_MCP_COMPACT", off);
-            assert!(!mcp_compact(), "{off:?} must disable compact");
-        }
-        for on in ["1", "true", "yes", "anything"] {
-            std::env::set_var("CARTOG_MCP_COMPACT", on);
-            assert!(mcp_compact(), "{on:?} must keep compact on");
-        }
-
-        match prev {
-            Some(v) => std::env::set_var("CARTOG_MCP_COMPACT", v),
-            None => std::env::remove_var("CARTOG_MCP_COMPACT"),
-        }
-    }
-
-    #[test]
-    fn rag_snippet_bounds_body_length() {
-        // The MCP rag_search default snips bodies via rag::search::snippet.
-        let long = "y".repeat(rag::search::SNIPPET_MAX_BYTES * 3);
-        let s = rag::search::snippet(&long);
-        assert!(s.len() <= rag::search::SNIPPET_MAX_BYTES);
-    }
-
-    // ── discover_plugin_pin tests (cartog_update arms the pin) ──
-    // Serialized via SERIAL because they mutate process-global env vars.
-
-    #[test]
-    fn discover_plugin_pin_reads_explicit_manifest() {
-        let _g = test_validate_call_counter::SERIAL.blocking_lock();
-        let prev_json = std::env::var_os("CARTOG_PLUGIN_JSON");
-        let prev_root = std::env::var_os("CLAUDE_PLUGIN_ROOT");
-        std::env::remove_var("CLAUDE_PLUGIN_ROOT");
-
-        let dir = tempfile::TempDir::new().unwrap();
-        let manifest = dir.path().join("plugin.json");
-        std::fs::write(&manifest, r#"{"name":"cartog","version":"0.20.0"}"#).unwrap();
-        std::env::set_var("CARTOG_PLUGIN_JSON", &manifest);
-        assert_eq!(discover_plugin_pin().as_deref(), Some("0.20.0"));
-
-        // Malformed (non-bare) version → None (fall back to latest).
-        std::fs::write(&manifest, r#"{"version":"v0.20.0"}"#).unwrap();
-        assert_eq!(
-            discover_plugin_pin(),
-            None,
-            "non-bare-semver pin must be rejected"
-        );
-
-        // No manifest discoverable → None.
-        std::env::remove_var("CARTOG_PLUGIN_JSON");
-        assert_eq!(discover_plugin_pin(), None);
-
-        match prev_json {
-            Some(v) => std::env::set_var("CARTOG_PLUGIN_JSON", v),
-            None => std::env::remove_var("CARTOG_PLUGIN_JSON"),
-        }
-        if let Some(v) = prev_root {
-            std::env::set_var("CLAUDE_PLUGIN_ROOT", v);
-        }
-    }
-
-    #[test]
-    fn discover_plugin_pin_reads_claude_plugin_root() {
-        let _g = test_validate_call_counter::SERIAL.blocking_lock();
-        let prev_json = std::env::var_os("CARTOG_PLUGIN_JSON");
-        let prev_root = std::env::var_os("CLAUDE_PLUGIN_ROOT");
-        std::env::remove_var("CARTOG_PLUGIN_JSON");
-
-        let dir = tempfile::TempDir::new().unwrap();
-        std::fs::create_dir_all(dir.path().join(".claude-plugin")).unwrap();
-        std::fs::write(
-            dir.path().join(".claude-plugin").join("plugin.json"),
-            r#"{"version":"0.21.0"}"#,
-        )
-        .unwrap();
-        std::env::set_var("CLAUDE_PLUGIN_ROOT", dir.path());
-        assert_eq!(discover_plugin_pin().as_deref(), Some("0.21.0"));
-
-        match prev_json {
-            Some(v) => std::env::set_var("CARTOG_PLUGIN_JSON", v),
-            None => std::env::remove_var("CARTOG_PLUGIN_JSON"),
-        }
-        match prev_root {
-            Some(v) => std::env::set_var("CLAUDE_PLUGIN_ROOT", v),
-            None => std::env::remove_var("CLAUDE_PLUGIN_ROOT"),
-        }
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn parse_arm_output_armed_maps_fields() {
-        let r = parse_arm_output(&output_ok(
-            r#"{"status":"armed","current":"0.19.0","target":"0.20.0","apply":"session-end-or-restart"}"#,
-            "",
-        ));
-        assert_eq!(r.status, "armed");
-        assert_eq!(r.target.as_deref(), Some("0.20.0"));
-        assert_eq!(r.apply, "session-end-or-restart");
-        assert!(r.message.contains("session ends"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn parse_arm_output_up_to_date_maps() {
-        let r = parse_arm_output(&output_ok(
-            r#"{"status":"up-to-date","current":"0.19.0","latest":"0.19.0"}"#,
-            "",
-        ));
-        assert_eq!(r.status, "up-to-date");
-        assert_eq!(r.apply, "n/a");
-        assert!(r.target.is_none());
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn parse_arm_output_cargo_maps_to_cargo_refused() {
-        let r = parse_arm_output(&output_ok(
-            r#"{"status":"cargo","message":"cartog was installed via cargo. Run `cargo install cartog --force` to upgrade."}"#,
-            "",
-        ));
-        assert_eq!(r.status, "cargo-refused");
-        assert!(r.message.contains("cargo install cartog --force"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn parse_arm_output_foreign_status_echoes_message_as_error() {
-        let r = parse_arm_output(&output_ok(
-            r#"{"status":"fetch-failed","message":"GitHub API returned status 500"}"#,
-            "",
-        ));
-        assert_eq!(r.status, "error");
-        assert!(r.message.contains("GitHub API returned status 500"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn parse_arm_output_skips_log_lines_before_json() {
-        // A daily-update-check hint or any noise line before the JSON must not
-        // break the parse — the reverse scan finds the real object.
-        let r = parse_arm_output(&output_ok(
-            "cartog: a new version is available\n{\"status\":\"armed\",\"target\":\"0.20.0\"}\n",
-            "",
-        ));
-        assert_eq!(r.status, "armed");
-        assert_eq!(r.target.as_deref(), Some("0.20.0"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn parse_arm_output_ignores_trailing_bare_scalar() {
-        // A trailing bare scalar that is valid JSON must NOT be picked over the
-        // real status object earlier in the stream.
-        let r = parse_arm_output(&output_ok(
-            "{\"status\":\"armed\",\"target\":\"1.2.3\"}\n99\n",
-            "",
-        ));
-        assert_eq!(r.status, "armed", "bare scalar must be skipped");
-        assert_eq!(r.target.as_deref(), Some("1.2.3"));
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn parse_arm_output_empty_output_names_exit_and_next_action() {
-        // Child produced nothing (e.g. SIGKILL). The message must still give a
-        // next step rather than a dead-end empty string.
-        let out = std::process::Output {
-            status: std::process::Command::new("sh")
-                .args(["-c", "exit 9"])
-                .status()
-                .expect("run sh"),
-            stdout: Vec::new(),
-            stderr: Vec::new(),
-        };
-        let r = parse_arm_output(&out);
-        assert_eq!(r.status, "error");
-        assert!(
-            r.message.contains("exit 9"),
-            "must name the exit code: {}",
-            r.message
-        );
-        assert!(
-            r.message.contains("--defer") || r.message.contains("/cartog-install"),
-            "must name a next action: {}",
-            r.message
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn parse_arm_output_nonjson_stderr_surfaces_detail() {
-        let r = parse_arm_output(&output_ok("", "boom: something broke"));
-        assert_eq!(r.status, "error");
-        assert!(r.message.contains("boom: something broke"));
-    }
-
-    // ── Tool annotation tests ──
-
-    /// Tools without side effects advertise `read_only_hint = true`; the three
-    /// side-effecting tools advertise `false`. Two of those write the DB
-    /// (`cartog_index`, `cartog_rag_index`); `cartog_update` arms a machine-level
-    /// deferred update. Clients use the hint to skip approval prompts for safe
-    /// tools and flag the rest.
-    #[test]
-    fn tool_annotations_mark_read_only_correctly() {
-        let side_effecting = ["cartog_index", "cartog_rag_index", "cartog_update"];
-        let tools = CartogServer::tool_router().list_all();
-
-        assert!(!tools.is_empty(), "router exposes tools");
-        for tool in &tools {
-            let ann = tool
-                .annotations
-                .as_ref()
-                .unwrap_or_else(|| panic!("{} has no annotations", tool.name));
-            let expected_read_only = !side_effecting.contains(&tool.name.as_ref());
-            assert_eq!(
-                ann.read_only_hint,
-                Some(expected_read_only),
-                "{} read_only_hint",
-                tool.name
-            );
-        }
-    }
-
-    /// Every tool exposes a human-readable title for client tool pickers.
-    #[test]
-    fn every_tool_has_a_title() {
-        for tool in CartogServer::tool_router().list_all() {
-            let title = tool
-                .annotations
-                .as_ref()
-                .and_then(|a| a.title.as_deref())
-                .or(tool.title.as_deref());
-            assert!(
-                title.is_some_and(|t| !t.is_empty()),
-                "{} has no title",
-                tool.name
-            );
-        }
-    }
-
-    // ── Output schema / structured content tests ──
-
-    /// Every read tool advertises an output schema so schema-aware clients can
-    /// validate `structuredContent`. The two write tools have no typed output.
-    #[test]
-    fn read_tools_advertise_output_schemas() {
-        let writers = ["cartog_index", "cartog_rag_index"];
-        for tool in CartogServer::tool_router().list_all() {
-            let has_schema = tool.output_schema.is_some();
-            let expected = !writers.contains(&tool.name.as_ref());
-            assert_eq!(has_schema, expected, "{} output_schema", tool.name);
-        }
-    }
-
-    /// `schema_for_output` rejects non-object schemas, so list tools must wrap
-    /// their arrays. Building each schema proves the wrappers stay object-typed
-    /// (a regression here would panic the tool macro at startup).
-    #[test]
-    fn output_schemas_are_objects() {
-        output_schema_for::<SymbolList>();
-        output_schema_for::<EdgeList>();
-        output_schema_for::<RefList>();
-        output_schema_for::<ImpactList>();
-        output_schema_for::<HierarchyList>();
-        output_schema_for::<MapResult>();
-        output_schema_for::<StatsResult>();
-        output_schema_for::<cartog_core::ChangesResult>();
-        output_schema_for::<rag::search::HybridSearchResult>();
-    }
-
-    /// Non-standard integer formats (`uint32`, …) are stripped so strict client
-    /// validators don't warn, while the field stays typed as an integer.
-    #[test]
-    fn output_schema_strips_nonstandard_int_formats() {
-        let schema = output_schema_for::<SymbolList>();
-        let value = serde_json::Value::Object((*schema).clone());
-
-        fn collect_formats(v: &serde_json::Value, out: &mut Vec<String>) {
-            match v {
-                serde_json::Value::Object(map) => {
-                    if let Some(serde_json::Value::String(f)) = map.get("format") {
-                        out.push(f.clone());
-                    }
-                    map.values().for_each(|v| collect_formats(v, out));
-                }
-                serde_json::Value::Array(items) => {
-                    items.iter().for_each(|v| collect_formats(v, out));
-                }
-                _ => {}
-            }
-        }
-        let mut formats = Vec::new();
-        collect_formats(&value, &mut formats);
-        assert!(
-            !formats
-                .iter()
-                .any(|f| NONSTANDARD_INT_FORMATS.contains(&f.as_str())),
-            "non-standard int formats leaked: {formats:?}"
-        );
-
-        // The integer field survives, just without the bogus format.
-        let start_line = &value["$defs"]["Symbol"]["properties"]["start_line"];
-        assert_eq!(start_line["type"], "integer");
-        assert!(start_line.get("format").is_none());
-    }
-
-    fn populated_memory_db() -> Database {
-        let db = Database::open_memory().expect("in-memory DB");
-        db.upsert_file(&cartog_core::FileInfo {
-            path: "test.py".to_string(),
-            last_modified: 0.0,
-            hash: "h".to_string(),
-            language: "python".to_string(),
-            num_symbols: 1,
-        })
-        .expect("upsert file");
-        db
-    }
-
-    /// A full (under-budget) response mirrors its payload into
-    /// `structuredContent` while keeping the bare-array text block.
-    #[test]
-    fn tool_response_attaches_structured_content_under_budget() {
-        let db = populated_memory_db();
-        let symbols = db.search("anything", None, None, 30).expect("search");
-        let json = serde_json::to_string_pretty(&symbols).expect("json");
-        let structured = serde_json::to_value(SymbolList { results: symbols }).ok();
-
-        let result = tool_response(&db, json, structured, "cartog_search", None).expect("response");
-
-        let structured = result
-            .structured_content
-            .expect("structured content present");
-        assert!(
-            structured.get("results").is_some(),
-            "structured content is the object wrapper"
-        );
-        assert_eq!(result.content.len(), 1, "text block retained");
-    }
-
-    /// An over-budget response drops `structuredContent` so an oversized payload
-    /// can't bypass the size cap, and the text block carries a truncation notice.
-    #[test]
-    fn tool_response_drops_structured_content_when_truncated() {
-        let db = populated_memory_db();
-        // Exceed the default 64KB cap deterministically (no env mutation).
-        let big = "x".repeat(mcp_max_bytes() + 1024);
-        let json = format!("[\"{big}\"]");
-        let structured = Some(serde_json::json!({ "results": [] }));
-
-        let result = tool_response(&db, json, structured, "cartog_search", None).expect("response");
-
-        assert!(
-            result.structured_content.is_none(),
-            "structured content dropped on truncation"
-        );
-        let text = match &result.content.first().expect("content").raw {
-            RawContent::Text(t) => &t.text,
-            _ => panic!("expected text content"),
-        };
-        assert!(
-            text.contains("Response truncated"),
-            "truncation notice present"
-        );
-    }
-
-    /// A staleness banner must not push a truncated response over the cap: the
-    /// banner's bytes are reserved before truncation, so banner + body ≤ cap.
-    #[test]
-    fn tool_response_with_banner_stays_under_cap() {
-        let db = populated_memory_db();
-        let big = "x".repeat(mcp_max_bytes() + 1024);
-        let json = format!("[\"{big}\"]");
-        // rag_pending on a rag tool fires the longest banner.
-        let stale = Some(snap(9, 0, 0));
-        let result = tool_response(&db, json, None, "cartog_rag_search", stale).expect("response");
-        let text = match &result.content.first().expect("content").raw {
-            RawContent::Text(t) => &t.text,
-            _ => panic!("expected text content"),
-        };
-        assert!(text.starts_with("⚠️"), "banner present: {}", &text[..40]);
-        assert!(
-            text.len() <= mcp_max_bytes(),
-            "banner + body must stay under the {}-byte cap, got {}",
-            mcp_max_bytes(),
-            text.len()
-        );
-    }
-
-    /// The final clamp also covers the NON-truncated path: a body just under the
-    /// banner-adjusted budget, plus a banner and an appended suggestion, must
-    /// still end up ≤ the cap (suffixes aren't individually budgeted).
-    #[test]
-    fn tool_response_banner_plus_suffix_stays_under_cap() {
-        let db = populated_memory_db();
-        let cap = mcp_max_bytes();
-        // Body sized so banner + body alone is just under cap; the appended
-        // suggestion would push it over without the final clamp.
-        let payload = "y".repeat(cap - 200);
-        let json = format!("[\"{payload}\"]");
-        let stale = Some(snap(3, 0, 0));
-        let result = tool_response(&db, json, None, "cartog_rag_search", stale).expect("response");
-        let text = match &result.content.first().expect("content").raw {
-            RawContent::Text(t) => &t.text,
-            _ => panic!("expected text content"),
-        };
-        assert!(
-            text.len() <= cap,
-            "banner + body + suffix must stay under {cap}, got {}",
-            text.len()
-        );
-    }
-
-    /// The size cap counts the structured copy too: when the text fits on its own
-    /// but text + structuredContent would exceed the budget, structured is dropped
-    /// (the text block is kept intact, not truncated).
-    #[test]
-    fn tool_response_drops_structured_when_combined_exceeds_cap() {
-        let db = populated_memory_db();
-        // Text alone is just under the cap; the structured mirror pushes the
-        // combined size over it.
-        let budget = mcp_max_bytes();
-        let payload = "y".repeat(budget * 3 / 4);
-        let json = format!("[\"{payload}\"]");
-        let structured = Some(serde_json::json!({ "results": [payload.clone()] }));
-
-        assert!(json.len() <= budget, "text alone fits the cap");
-        let result = tool_response(&db, json, structured, "cartog_search", None).expect("response");
-
-        assert!(
-            result.structured_content.is_none(),
-            "structured dropped when combined size exceeds cap"
-        );
-        let text = match &result.content.first().expect("content").raw {
-            RawContent::Text(t) => &t.text,
-            _ => panic!("expected text content"),
-        };
-        assert!(
-            !text.contains("Response truncated"),
-            "text block fits on its own, so it is not truncated"
-        );
-    }
-
-    /// `StatsResult` flattens `IndexStats` and adds role + watcher fields at the
-    /// top level (no nested `stats` object), matching the documented shape.
-    #[test]
-    fn stats_result_flattens_index_stats() {
-        let db = populated_memory_db();
-        let stats = db.stats().expect("stats");
-        let result = StatsResult {
-            stats,
-            role: Role::ReadOnly,
-            watcher_active: false,
-        };
-        let value = serde_json::to_value(&result).expect("serialize");
-        let obj = value.as_object().expect("object");
-        assert!(obj.contains_key("num_files"), "flattened stats field");
-        // Role serializes to the exact wire string (hyphen preserved).
-        assert_eq!(obj.get("role").and_then(|v| v.as_str()), Some("read-only"));
-        assert_eq!(
-            obj.get("watcher_active").and_then(|v| v.as_bool()),
-            Some(false)
-        );
-        assert!(!obj.contains_key("stats"), "stats is flattened, not nested");
-    }
-
-    // ── Path validation tests ──
-
-    #[test]
-    fn validate_path_dot_is_allowed() {
-        let result = validate_path_within_cwd(".");
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn validate_path_subdirectory_is_allowed() {
-        let result = validate_path_within_cwd("src");
-        // May not exist in test env, but should not be rejected as "outside CWD"
-        // (normalize_path handles non-existent paths)
-        assert!(result.is_ok() || result.unwrap_err().contains("cannot resolve"));
-    }
-
-    #[test]
-    fn validate_path_parent_escape_is_rejected() {
-        let result = validate_path_within_cwd("../../etc/passwd");
-        assert!(result.is_err());
-        assert!(
-            result
-                .unwrap_err()
-                .contains("outside the project directory"),
-            "should reject path traversal"
-        );
-    }
-
-    #[test]
-    fn validate_path_absolute_outside_cwd_is_rejected() {
-        let result = validate_path_within_cwd("/etc/passwd");
-        assert!(result.is_err());
-        assert!(
-            result
-                .unwrap_err()
-                .contains("outside the project directory"),
-            "should reject absolute paths outside CWD"
-        );
-    }
-
-    #[test]
-    fn validate_path_absolute_inside_cwd_is_allowed() {
-        let cwd = std::env::current_dir().expect("CWD");
-        let inside = cwd.join("src");
-        let result = validate_path_within_cwd(inside.to_str().expect("utf-8 path"));
-        // src/ exists in this project
-        assert!(result.is_ok());
-    }
-
-    #[test]
-    fn validate_path_dotdot_in_middle_is_rejected() {
-        let result = validate_path_within_cwd("src/../../etc");
-        assert!(result.is_err());
-    }
-
-    // ── Normalize path tests ──
-
-    #[test]
-    fn normalize_removes_dot() {
-        let p = normalize_path(Path::new("/a/./b/./c"));
-        assert_eq!(p, PathBuf::from("/a/b/c"));
-    }
-
-    #[test]
-    fn normalize_resolves_parent() {
-        let p = normalize_path(Path::new("/a/b/../c"));
-        assert_eq!(p, PathBuf::from("/a/c"));
-    }
-
-    // ── Depth capping ──
-
-    /// Verify depth is clamped at MAX_IMPACT_DEPTH.
-    #[test]
-    fn impact_depth_is_capped() {
-        fn resolve_depth(input: Option<u32>) -> u32 {
-            input.unwrap_or(3).min(MAX_IMPACT_DEPTH)
-        }
-        assert_eq!(resolve_depth(Some(999)), MAX_IMPACT_DEPTH);
-        assert_eq!(resolve_depth(Some(5)), 5);
-    }
-
-    /// Verify default depth when None is provided.
-    #[test]
-    fn impact_depth_default() {
-        fn resolve_depth(input: Option<u32>) -> u32 {
-            input.unwrap_or(3).min(MAX_IMPACT_DEPTH)
-        }
-        assert_eq!(resolve_depth(None), 3);
-    }
-
-    // ── Edge kind parsing ──
-
-    #[test]
-    fn parse_valid_edge_kinds() {
-        assert_eq!("calls".parse::<EdgeKind>().unwrap(), EdgeKind::Calls);
-        assert_eq!("imports".parse::<EdgeKind>().unwrap(), EdgeKind::Imports);
-        assert_eq!("inherits".parse::<EdgeKind>().unwrap(), EdgeKind::Inherits);
-        assert_eq!(
-            "references".parse::<EdgeKind>().unwrap(),
-            EdgeKind::References
-        );
-        assert_eq!("raises".parse::<EdgeKind>().unwrap(), EdgeKind::Raises);
-    }
-
-    #[test]
-    fn parse_invalid_edge_kind_fails() {
-        assert!("invalid".parse::<EdgeKind>().is_err());
-        assert!("CALLS".parse::<EdgeKind>().is_err());
-        assert!("".parse::<EdgeKind>().is_err());
-    }
-
-    // ── Edge provenance in structured output ──
-
-    #[test]
-    fn json_output_includes_provenance_when_present() {
-        let mut edge = cartog_core::Edge::new("s:1", "foo", EdgeKind::Calls, "a.py", 1);
-        edge.provenance = Some(cartog_core::EdgeProvenance::SameFile);
-        let value = serde_json::to_value(EdgeList {
-            results: vec![edge],
-        })
-        .unwrap();
-        assert_eq!(value["results"][0]["provenance"], "same_file");
-    }
-
-    #[test]
-    fn json_output_omits_provenance_when_absent() {
-        // A freshly extracted edge has no provenance; skip_serializing_if drops
-        // the key entirely so the wire format stays clean.
-        let edge = cartog_core::Edge::new("s:1", "foo", EdgeKind::Calls, "a.py", 1);
-        let value = serde_json::to_value(EdgeList {
-            results: vec![edge],
-        })
-        .unwrap();
-        assert!(value["results"][0].get("provenance").is_none());
-    }
-
-    // ── Tool handler tests (using in-memory DB) ──
-
-    // These test the underlying DB operations that the MCP handlers call.
-    // We cannot easily construct MCP tool calls in unit tests without a full
-    // server, so we test the DB layer directly with the same patterns.
-
-    #[test]
-    fn empty_db_outline_returns_empty() {
-        let db = Database::open_memory().expect("in-memory DB");
-        let result = db.outline("nonexistent.py").expect("query");
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn empty_db_refs_returns_empty() {
-        let db = Database::open_memory().expect("in-memory DB");
-        let result = db.refs("nonexistent", None).expect("query");
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn empty_db_callees_returns_empty() {
-        let db = Database::open_memory().expect("in-memory DB");
-        let result = db.callees("nonexistent").expect("query");
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn empty_db_impact_returns_empty() {
-        let db = Database::open_memory().expect("in-memory DB");
-        let result = db.impact("nonexistent", 3).expect("query");
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn empty_db_hierarchy_returns_empty() {
-        let db = Database::open_memory().expect("in-memory DB");
-        let result = db.hierarchy("nonexistent").expect("query");
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn empty_db_deps_returns_empty() {
-        let db = Database::open_memory().expect("in-memory DB");
-        let result = db.file_deps("nonexistent.py").expect("query");
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn empty_db_search_returns_empty() {
-        let db = Database::open_memory().expect("in-memory DB");
-        let result = db.search("foo", None, None, 20).expect("query");
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn did_you_mean_suffix_lists_candidates() {
-        let cands = vec!["ReviewResult".to_string(), "ReviewComment".to_string()];
-        let suffix = did_you_mean_suffix("Revie", &cands).expect("suffix");
-        assert!(suffix.contains("Did you mean: ReviewResult, ReviewComment"));
-        assert!(suffix.contains("cartog_search"));
-    }
-
-    #[test]
-    fn did_you_mean_suffix_none_on_exact_match() {
-        let cands = vec!["ReviewResult".to_string()];
-        assert!(did_you_mean_suffix("ReviewResult", &cands).is_none());
-    }
-
-    #[test]
-    fn did_you_mean_suffix_none_without_candidates() {
-        assert!(did_you_mean_suffix("Whatever", &[]).is_none());
-    }
-
-    fn snap(rag_pending: u32, change_seq: u64, reindexed_seq: u64) -> StaleSnapshot {
-        StaleSnapshot {
-            rag_pending,
-            change_seq,
-            reindexed_seq,
-        }
-    }
-
-    #[test]
-    fn stale_banner_none_when_not_stale() {
-        assert!(stale_banner(Some(snap(0, 10, 10)), "cartog_rag_search").is_none());
-        assert!(stale_banner(None, "cartog_rag_search").is_none());
-    }
-
-    #[test]
-    fn stale_banner_rag_only_warns_semantic_tools() {
-        // Pending embeddings warn rag_search/context...
-        assert!(stale_banner(Some(snap(3, 10, 10)), "cartog_rag_search").is_some());
-        assert!(stale_banner(Some(snap(3, 10, 10)), "cartog_context").is_some());
-        // ...but not a structural tool (no debounce gap here).
-        assert!(stale_banner(Some(snap(3, 10, 10)), "cartog_refs").is_none());
-    }
-
-    #[test]
-    fn stale_banner_structural_warns_every_read_tool() {
-        // A change after the last reindex warns refs and rag_search alike.
-        assert!(stale_banner(Some(snap(0, 20, 10)), "cartog_refs").is_some());
-        assert!(stale_banner(Some(snap(0, 20, 10)), "cartog_rag_search").is_some());
-    }
-
-    #[test]
-    fn search_limit_is_capped() {
-        assert_eq!(999u32.min(MAX_SEARCH_LIMIT), MAX_SEARCH_LIMIT);
-        assert_eq!(30u32.min(MAX_SEARCH_LIMIT), 30);
-    }
-
-    #[test]
-    fn empty_db_stats_returns_zeros() {
-        let db = Database::open_memory().expect("in-memory DB");
-        let stats = db.stats().expect("query");
-        assert_eq!(stats.num_files, 0);
-        assert_eq!(stats.num_symbols, 0);
-        assert_eq!(stats.num_edges, 0);
-        assert_eq!(stats.num_resolved, 0);
-    }
-
-    // ── Response serialization tests ──
-
-    #[test]
-    fn ref_entry_serializes() {
-        let entry = RefEntry {
-            edge: cartog_core::Edge::new("src:foo:1", "bar", EdgeKind::Calls, "src/main.py", 10),
-            source: None,
-        };
-        let json = serde_json::to_string(&entry).expect("serialize");
-        assert!(json.contains("\"bar\""));
-        assert!(json.contains("\"calls\""));
-    }
-
-    #[test]
-    fn impact_entry_serializes() {
-        let entry = ImpactEntry {
-            edge: cartog_core::Edge::new("src:foo:1", "bar", EdgeKind::Calls, "src/main.py", 10),
-            depth: 2,
-        };
-        let json = serde_json::to_string(&entry).expect("serialize");
-        assert!(json.contains("\"depth\":2"));
-    }
-
-    #[test]
-    fn hierarchy_entry_serializes() {
-        let entry = HierarchyEntry {
-            child: "Dog".to_string(),
-            parent: "Animal".to_string(),
-        };
-        let json = serde_json::to_string(&entry).expect("serialize");
-        assert!(json.contains("\"Dog\""));
-        assert!(json.contains("\"Animal\""));
-    }
-
-    // ── PID-file lock tests ──
-
-    #[test]
-    fn pid_file_acquired_when_lock_dir_set() {
-        let _guard = env_mutex().lock().unwrap_or_else(|e| e.into_inner());
-        let dir = tempfile::TempDir::new().unwrap();
-        let opts = ServerOptions {
-            pid_lock_dir: Some(dir.path().to_path_buf()),
-            pid_lock_slot: Some(SERVE_LOCK_SLOT.to_string()),
-        };
-        let outcome = acquire_serve_lock(&opts).expect("acquire");
-        let lock = match outcome {
-            ServeLockOutcome::Primary(l) => l,
-            other => panic!("expected Primary, got {other:?}"),
-        };
-        let path = dir.path().join(format!("{SERVE_LOCK_SLOT}.pid"));
-        assert!(path.exists(), "PID file should exist while lock is held");
-        // File is now two lines (pid + start_time); only the first line is the PID.
-        let contents = std::fs::read_to_string(&path).unwrap();
-        let pid: u32 = contents.lines().next().unwrap().trim().parse().unwrap();
-        assert_eq!(pid, std::process::id());
-        drop(lock);
-        assert!(
-            !path.exists(),
-            "PID file should be removed once the lock is dropped"
-        );
-    }
-
-    #[test]
-    fn pid_file_skipped_when_lock_dir_unset() {
-        let opts = ServerOptions::default();
-        let outcome = acquire_serve_lock(&opts).expect("noop");
-        assert!(
-            matches!(outcome, ServeLockOutcome::Untracked),
-            "no lock dir → Untracked"
-        );
-    }
-
-    /// Serialize tests that read or mutate the `CARTOG_SINGLE_WRITER` env
-    /// var. The variable is process-global; cargo test runs cases in
-    /// parallel by default, so without this mutex a concurrent setter
-    /// flips the value mid-read on another thread.
-    fn env_mutex() -> &'static std::sync::Mutex<()> {
-        static M: std::sync::OnceLock<std::sync::Mutex<()>> = std::sync::OnceLock::new();
-        M.get_or_init(|| std::sync::Mutex::new(()))
-    }
-
-    #[test]
-    fn acquire_serve_lock_rejects_dir_without_slot() {
-        // Regression: a half-configured ServerOptions (pid_lock_dir set,
-        // pid_lock_slot None) used to silently fall back to the global
-        // SERVE_LOCK_SLOT, letting an embedder collide with — or be hidden
-        // from — a CLI peer that derives a DB-scoped slot. The mixed-scope
-        // hazard must surface as a hard error.
-        let _guard = env_mutex().lock().unwrap_or_else(|e| e.into_inner());
-        let dir = tempfile::TempDir::new().unwrap();
-        let opts = ServerOptions {
-            pid_lock_dir: Some(dir.path().to_path_buf()),
-            pid_lock_slot: None,
-        };
-        let err = acquire_serve_lock(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("pid_lock_slot is None"),
-            "error must explain the misconfiguration, got: {err}"
-        );
-    }
-
-    #[test]
-    fn acquire_serve_lock_rejects_slot_without_dir() {
-        // Inverse half-config: pid_lock_slot set but pid_lock_dir is
-        // None. Pre-fix the slot was silently ignored and the function
-        // returned Ok(Untracked), losing the caller's intent. Must
-        // surface as a hard error.
-        let _guard = env_mutex().lock().unwrap_or_else(|e| e.into_inner());
-        let opts = ServerOptions {
-            pid_lock_dir: None,
-            pid_lock_slot: Some("serve-deadbeef".to_string()),
-        };
-        let err = acquire_serve_lock(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("pid_lock_dir is None"),
-            "error must explain the misconfiguration, got: {err}"
-        );
-    }
-
-    #[test]
-    fn distinct_slots_for_different_dbs_do_not_collide() {
-        // Two cartog peers serving different DBs in the same per-user state
-        // dir must coexist. Pre-PR, both fought over a single `serve.pid`
-        // slot; with DB-scoped slots they each claim their own
-        // `serve-<hash>.pid`.
-        let _guard = env_mutex().lock().unwrap_or_else(|e| e.into_inner());
-        let dir = tempfile::TempDir::new().unwrap();
-        let opts_a = ServerOptions {
-            pid_lock_dir: Some(dir.path().to_path_buf()),
-            pid_lock_slot: Some("serve-aaaa1111".to_string()),
-        };
-        let opts_b = ServerOptions {
-            pid_lock_dir: Some(dir.path().to_path_buf()),
-            pid_lock_slot: Some("serve-bbbb2222".to_string()),
-        };
-        let _a = match acquire_serve_lock(&opts_a).expect("acquire A") {
-            ServeLockOutcome::Primary(l) => l,
-            other => panic!("expected Primary for A, got {other:?}"),
-        };
-        let _b = match acquire_serve_lock(&opts_b).expect("acquire B") {
-            ServeLockOutcome::Primary(l) => l,
-            other => panic!("expected Primary for B, got {other:?}"),
-        };
-        // Both PID files present on disk, no Held collision.
-        assert!(dir.path().join("serve-aaaa1111.pid").exists());
-        assert!(dir.path().join("serve-bbbb2222.pid").exists());
-    }
-
-    #[test]
-    fn serve_to_watch_slot_preserves_db_scope() {
-        // The watcher slot is derived from the serve slot so both PID files
-        // for the same DB share their scope suffix.
-        assert_eq!(serve_to_watch_slot("serve").unwrap(), "watch");
-        assert_eq!(serve_to_watch_slot("serve-abc123").unwrap(), "watch-abc123");
-        // Off-pattern inputs that start with the bytes "serve" but are NOT
-        // a serve-family slot must be REJECTED — silently folding them to
-        // the global watch slot would let distinct embedders collide on
-        // `watch.pid` while their serve slots stay distinct.
-        for bad in [
-            "unknown-prefix",
-            "server",
-            "serverless",
-            "servefoo",
-            "Serve",
-            "",
-            "serve-", // trailing-dash with empty hex
-        ] {
-            assert!(
-                serve_to_watch_slot(bad).is_err(),
-                "expected off-pattern slot {bad:?} to be rejected"
-            );
-        }
-    }
-
-    proptest::proptest! {
-        /// `serve-<nonempty>` maps to `watch-<same suffix>`, preserving the DB
-        /// scope verbatim.
-        #[test]
-        fn serve_to_watch_slot_round_trips_suffix(suffix in "[0-9a-f]{1,16}") {
-            let got = serve_to_watch_slot(&format!("serve-{suffix}")).unwrap();
-            proptest::prop_assert_eq!(got, format!("watch-{suffix}"));
-        }
-
-        /// Total contract over arbitrary input, checked against the OUTPUT rather
-        /// than a re-implemented accept rule: the only accepted slots are `serve`
-        /// (→ `watch`) and `serve-<nonempty>` (→ `watch-<same>`); every other
-        /// input is rejected. Folding an off-pattern slot to the global watch slot
-        /// would let distinct embedders collide on `watch.pid`.
-        #[test]
-        fn serve_to_watch_slot_contract(s in ".{0,20}") {
-            match serve_to_watch_slot(&s) {
-                Ok(out) if s == "serve" => proptest::prop_assert_eq!(out, "watch"),
-                Ok(out) => {
-                    // The only other accepted form: the suffix is carried verbatim.
-                    let suffix = s.strip_prefix("serve-");
-                    proptest::prop_assert!(
-                        suffix.is_some_and(|r| !r.is_empty()),
-                        "accepted off-pattern {s:?}"
-                    );
-                    proptest::prop_assert_eq!(out, format!("watch-{}", suffix.unwrap()));
-                }
-                Err(_) => {
-                    proptest::prop_assert_ne!(&s, "serve");
-                    proptest::prop_assert!(
-                        s.strip_prefix("serve-").map_or(true, str::is_empty),
-                        "rejected a valid serve-<nonempty> slot: {s:?}"
-                    );
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn second_acquire_for_same_dir_reports_held() {
-        // Two acquire_serve_lock calls against the same dir: the first wins,
-        // the second must surface Held(_) with the first's PID so the caller
-        // can branch.
-        let _guard = env_mutex().lock().unwrap_or_else(|e| e.into_inner());
-        let dir = tempfile::TempDir::new().unwrap();
-        let opts = ServerOptions {
-            pid_lock_dir: Some(dir.path().to_path_buf()),
-            pid_lock_slot: Some(SERVE_LOCK_SLOT.to_string()),
-        };
-        let _first = match acquire_serve_lock(&opts).expect("first acquire") {
-            ServeLockOutcome::Primary(l) => l,
-            other => panic!("expected Primary, got {other:?}"),
-        };
-        let second = acquire_serve_lock(&opts).expect("second acquire returns ok");
-        match second {
-            ServeLockOutcome::Held(held) => {
-                assert_eq!(held.slot, SERVE_LOCK_SLOT);
-                assert_eq!(held.pid, std::process::id());
-            }
-            other => panic!("expected Held, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn kill_switch_disables_election() {
-        let _guard = env_mutex().lock().unwrap_or_else(|e| e.into_inner());
-        // CARTOG_SINGLE_WRITER=0 must let a second acquire_overwriting-style
-        // call succeed despite a live first holder. Restoring the env var
-        // afterwards is best-effort; tests in a single binary share env so
-        // we set + unset around the call site.
-        let dir = tempfile::TempDir::new().unwrap();
-        let opts = ServerOptions {
-            pid_lock_dir: Some(dir.path().to_path_buf()),
-            pid_lock_slot: Some(SERVE_LOCK_SLOT.to_string()),
-        };
-        let _first = match acquire_serve_lock(&opts).expect("first acquire") {
-            ServeLockOutcome::Primary(l) => l,
-            other => panic!("expected Primary, got {other:?}"),
-        };
-
-        // SAFETY: tests in `cargo test` run in threads, but env mutation is
-        // process-global. Other tests in this file don't depend on this var,
-        // and we restore it before returning.
-        let prev = std::env::var(SINGLE_WRITER_ENV).ok();
-        // SAFETY: env vars are inherently process-wide and tests share them.
-        // We restore the prior value before this test returns so adjacent
-        // tests aren't affected.
-        unsafe {
-            std::env::set_var(SINGLE_WRITER_ENV, "0");
-        }
-        let result = acquire_serve_lock(&opts);
-        // SAFETY: same reason — restoring prior state regardless of outcome.
-        unsafe {
-            match prev {
-                Some(v) => std::env::set_var(SINGLE_WRITER_ENV, v),
-                None => std::env::remove_var(SINGLE_WRITER_ENV),
-            }
-        }
-        match result.expect("kill switch acquire") {
-            ServeLockOutcome::Primary(_) => {} // expected
-            other => panic!("expected Primary with kill switch, got {other:?}"),
-        }
-    }
-
-    // ── Role / read-only attach tests (Phase 4) ──
-
-    /// Deterministic in-memory embedding provider for tests, sized to the
-    /// default embedding dimension so the DB schema matches. Avoids loading
-    /// the real ONNX model (which may be absent in CI coverage runners).
-    fn test_provider() -> Box<dyn rag::provider::EmbeddingProvider> {
-        Box::new(rag::provider::test_utils::MockEmbeddingProvider::new(
-            rag::EMBEDDING_DIM,
-        ))
-    }
-
-    #[test]
-    fn primary_server_reports_primary_role() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let server = CartogServer::new_with_provider(
-            &db_path,
-            test_provider(),
-            indexer::RedactionConfig::disabled(),
-            indexer::WalkFilter::unrestricted(),
-            Role::Primary,
-        )
-        .expect("primary server constructs");
-        assert_eq!(server.role(), Role::Primary);
-    }
-
-    #[test]
-    fn read_only_server_reports_read_only_role() {
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        // First open writable to materialize the file with current schema.
-        {
-            let _primary = CartogServer::new_with_provider(
-                &db_path,
-                test_provider(),
-                indexer::RedactionConfig::disabled(),
-                indexer::WalkFilter::unrestricted(),
-                Role::Primary,
-            )
-            .expect("primary server constructs");
-        }
-        let reader = CartogServer::new_with_provider(
-            &db_path,
-            test_provider(),
-            indexer::RedactionConfig::disabled(),
-            indexer::WalkFilter::unrestricted(),
-            Role::ReadOnly,
-        )
-        .expect("read-only server constructs");
-        assert_eq!(reader.role(), Role::ReadOnly);
-    }
-
-    #[test]
-    fn promoter_validate_pinned_state_matches_when_unchanged() {
-        let _serial = test_validate_call_counter::SERIAL.blocking_lock();
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        {
-            // Materialize a DB so open_readonly can later read its state.
-            let _primary = Database::open(&db_path, 384).unwrap();
-        }
-        let pinned = Database::open_readonly(&db_path)
-            .unwrap()
-            .pinned_attach()
-            .cloned();
-        validate_pinned_state(&db_path, pinned.as_ref()).expect("matching pin must validate");
-    }
-
-    #[test]
-    fn promoter_validate_pinned_state_detects_schema_bump() {
-        let _serial = test_validate_call_counter::SERIAL.blocking_lock();
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let pinned = {
-            let db = Database::open(&db_path, 384).unwrap();
-            drop(db);
-            Database::open_readonly(&db_path)
-                .unwrap()
-                .pinned_attach()
-                .cloned()
-        };
-        // Simulate another writer upgrading the schema underneath us.
-        {
-            let db = Database::open(&db_path, 384).unwrap();
-            db.set_metadata("schema_version", "9999").unwrap();
-        }
-        let result = validate_pinned_state(&db_path, pinned.as_ref());
-        // open_readonly returns SchemaDrift; validate_pinned_state wraps as anyhow.
-        assert!(result.is_err(), "schema bump under us must fail validation");
-    }
-
-    #[test]
-    fn atomic_role_round_trip() {
-        let r = AtomicRole::new(Role::ReadOnly);
-        assert_eq!(r.load(), Role::ReadOnly);
-        r.store(Role::Primary);
-        assert_eq!(r.load(), Role::Primary);
-        r.store(Role::ReadOnly);
-        assert_eq!(r.load(), Role::ReadOnly);
-    }
-
-    #[test]
-    fn read_only_server_refuses_write_tools() {
-        // refuse_if_read_only is the helper gating cartog_index and
-        // cartog_rag_index. Verify both call sites get an error in
-        // ReadOnly mode and pass through silently in Primary mode.
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        {
-            let _primary = CartogServer::new_with_provider(
-                &db_path,
-                test_provider(),
-                indexer::RedactionConfig::disabled(),
-                indexer::WalkFilter::unrestricted(),
-                Role::Primary,
-            )
-            .expect("primary server constructs");
-        }
-        let reader = CartogServer::new_with_provider(
-            &db_path,
-            test_provider(),
-            indexer::RedactionConfig::disabled(),
-            indexer::WalkFilter::unrestricted(),
-            Role::ReadOnly,
-        )
-        .expect("read-only server constructs");
-
-        let err = reader
-            .refuse_if_read_only("cartog_index")
-            .expect("read-only must refuse");
-        let msg = format!("{err:?}");
-        assert!(
-            msg.contains("read-only") && msg.contains("cartog_index"),
-            "error must name the gate and the tool, got: {msg}"
-        );
-        // cartog_index → suggests `cartog index` (graph), not `cartog rag index`.
-        assert!(
-            msg.contains("cartog index") && !msg.contains("cartog rag index"),
-            "cartog_index refusal must suggest `cartog index`, got: {msg}"
-        );
-        // Drops the misleading "~5s" promise.
-        assert!(
-            !msg.contains("~5s"),
-            "refusal must not promise an exact pickup latency, got: {msg}"
-        );
-
-        let err = reader
-            .refuse_if_read_only("cartog_rag_index")
-            .expect("read-only must refuse");
-        let msg = format!("{err:?}");
-        assert!(
-            msg.contains("read-only") && msg.contains("cartog_rag_index"),
-            "error must name the gate and the tool, got: {msg}"
-        );
-        // cartog_rag_index → suggests `cartog rag index` (vectors).
-        assert!(
-            msg.contains("cartog rag index"),
-            "cartog_rag_index refusal must suggest `cartog rag index`, got: {msg}"
-        );
-
-        let primary = CartogServer::new_with_provider(
-            &db_path,
-            test_provider(),
-            indexer::RedactionConfig::disabled(),
-            indexer::WalkFilter::unrestricted(),
-            Role::Primary,
-        )
-        .expect("primary reconstructs");
-        assert!(
-            primary.refuse_if_read_only("cartog_index").is_none(),
-            "primary must NOT refuse"
-        );
-    }
-
-    #[test]
-    fn cartog_update_is_registered_and_not_gated_read_only() {
-        // cartog_update arms a machine-level deferred update, not a DB write,
-        // so it must be available even on a read-only secondary. The router
-        // lists it regardless of role, and the handler never consults
-        // refuse_if_read_only for it.
-        let names: Vec<String> = CartogServer::tool_router()
-            .list_all()
-            .into_iter()
-            .map(|t| t.name.to_string())
-            .collect();
-        assert!(
-            names.iter().any(|n| n == "cartog_update"),
-            "cartog_update must be registered, got: {names:?}"
-        );
-
-        let writers = ["cartog_index", "cartog_rag_index"];
-        assert!(
-            !writers.contains(&"cartog_update"),
-            "cartog_update must NOT be in the DB-write set that refuse_if_read_only gates"
-        );
-    }
-
-    // ── Index + LSP gate tests ──
-
-    #[cfg(feature = "lsp")]
-    #[test]
-    fn index_with_optional_lsp_skips_lsp_on_noop_reindex() {
-        // Regression guard for the MCP-side gate: when no file changed since
-        // the previous index, the LSP pass MUST be skipped — otherwise we
-        // re-spawn rust-analyzer / pyright on every cartog_index call.
-        //
-        // Copies the auth fixture to a tempdir so a real source edit can be
-        // applied between calls without touching the repo.
-        use cartog_db::Database;
-        use cartog_lsp::manager::LspManager;
-
-        let fixtures_src = Path::new(env!("CARGO_MANIFEST_DIR")).join("../../tests/fixtures/auth");
-        if !fixtures_src.exists() {
-            return;
-        }
-
-        let tmp = tempfile::tempdir().unwrap();
-        let fixtures = tmp.path().join("auth");
-        std::fs::create_dir_all(&fixtures).unwrap();
-        for entry in std::fs::read_dir(&fixtures_src).unwrap() {
-            let entry = entry.unwrap();
-            std::fs::copy(entry.path(), fixtures.join(entry.file_name())).unwrap();
-        }
-
-        let db = Arc::new(Mutex::new(Database::open_memory().unwrap()));
-        let lsp_mgr = Arc::new(Mutex::new(LspManager::new(&fixtures)));
-
-        // First call primes the index. dirty_files > 0 → LSP is allowed (it may
-        // resolve nothing if pyright isn't on PATH, but the gate must let it run).
-        let r1 = index_with_optional_lsp(
-            &db,
-            &lsp_mgr,
-            &fixtures,
-            false,
-            None,
-            None,
-            indexer::RedactionConfig::disabled(),
-            &indexer::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-        assert!(
-            r1.dirty_files > 0,
-            "first index must report dirty files (got {})",
-            r1.dirty_files
-        );
-
-        // Second call without changes must be a no-op AND must skip LSP.
-        let r2 = index_with_optional_lsp(
-            &db,
-            &lsp_mgr,
-            &fixtures,
-            false,
-            None,
-            None,
-            indexer::RedactionConfig::disabled(),
-            &indexer::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-        assert_eq!(r2.dirty_files, 0);
-        assert_eq!(
-            r2.edges_lsp_resolved, 0,
-            "no-op reindex must skip LSP (MCP-side gate broken)"
-        );
-        assert_eq!(
-            r2.edges_marked_external, 0,
-            "no-op reindex must not produce new external marks"
-        );
-        assert_eq!(r2.files_indexed, 0);
-    }
-
-    /// Tempdir project with one guaranteed-unresolvable Python call, indexed
-    /// with `lsp=false` so its edge is sealed at state=4.
-    #[cfg(feature = "lsp")]
-    fn sealed_py_fixture() -> (
-        tempfile::TempDir,
-        std::path::PathBuf,
-        Arc<Mutex<cartog_db::Database>>,
-    ) {
-        let tmp = tempfile::tempdir().unwrap();
-        // Tempdirs are dot-prefixed on macOS; is_ignored() rejects a dotted walk root.
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir_all(&root).unwrap();
-        std::fs::write(root.join("a.py"), "def caller():\n    nowhere('x')\n").unwrap();
-
-        let db = cartog_db::Database::open_memory().unwrap();
-        indexer::index_directory(
-            &db,
-            &root,
-            false,
-            false,
-            None,
-            None,
-            indexer::RedactionConfig::disabled(),
-            &std::collections::HashMap::new(),
-            &indexer::WalkFilter::unrestricted(),
-        )
-        .unwrap();
-        assert!(
-            db.has_heuristic_exhausted().unwrap(),
-            "no-lsp index must seal the unresolvable edge"
-        );
-        (tmp, root, Arc::new(Mutex::new(db)))
-    }
-
-    /// Manager whose python server is a binary that cannot exist, so the start
-    /// fails deterministically regardless of the host PATH.
-    #[cfg(feature = "lsp")]
-    fn no_server_manager(root: &Path) -> cartog_lsp::manager::LspManager {
-        let overrides = std::collections::HashMap::from([(
-            "python".to_string(),
-            vec!["cartog-test-no-such-binary".to_string()],
-        )]);
-        cartog_lsp::manager::LspManager::with_overrides(root, overrides)
-    }
-
-    #[cfg(feature = "lsp")]
-    #[test]
-    fn warm_lsp_pass_reopens_and_reseals_without_servers() {
-        // Regression for the #114 seal: `index_directory(lsp=false)` seals
-        // unresolved edges at state=4 and the warm pass queries only state=0 —
-        // without the reopen it silently resolved nothing.
-        let (_tmp, root, db) = sealed_py_fixture();
-        let db = db.lock().unwrap();
-        let sealed = db.count_edges_in_state(4).unwrap();
-        let mut mgr = no_server_manager(&root);
-
-        let outcome = warm_lsp_pass(&db, &mut mgr, &root, None, None).unwrap();
-
-        assert_eq!(
-            outcome.reopened, sealed,
-            "seals must reopen before the pass"
-        );
-        assert!(!outcome.stats.any_server_started);
-        assert_eq!(outcome.resealed, sealed, "serverless pass must re-seal");
-        assert_eq!(db.count_edges_in_state(4).unwrap(), sealed);
-        assert_eq!(db.count_edges_in_state(0).unwrap(), 0);
-    }
-
-    #[cfg(feature = "lsp")]
-    #[test]
-    fn warm_lsp_pass_reseals_on_cancel() {
-        // A cancelled pass has no surrounding tx on the MCP path: the reopen
-        // must be undone or the no-op catch-up loses sight of the backlog.
-        let (_tmp, root, db) = sealed_py_fixture();
-        let db = db.lock().unwrap();
-        let sealed = db.count_edges_in_state(4).unwrap();
-        let mut mgr = no_server_manager(&root);
-
-        let tripped: indexer::CancelProbe<'_> = &|| true;
-        let err = warm_lsp_pass(&db, &mut mgr, &root, None, Some(tripped)).unwrap_err();
-
-        assert!(
-            format!("{err:?}").contains("cancelled"),
-            "cancel must surface as an error, got: {err:?}"
-        );
-        assert_eq!(
-            db.count_edges_in_state(4).unwrap(),
-            sealed,
-            "pre-pass seal must be restored on cancel"
-        );
-        assert_eq!(db.count_edges_in_state(0).unwrap(), 0);
-    }
-
-    #[cfg(feature = "lsp")]
-    #[test]
-    fn catch_up_runs_warm_pass_on_noop_index_with_sealed_backlog() {
-        let (_tmp, root, db) = sealed_py_fixture();
-        let latch = std::sync::atomic::AtomicBool::new(false);
-        let mgr = Arc::new(Mutex::new(no_server_manager(&root)));
-        let (tx, mut rx) = tokio::sync::mpsc::channel(64);
-
-        let noop = indexer::IndexResult::default();
-        let out = catch_up_lsp(&db, &mgr, &latch, &root, Some(tx), None, noop).unwrap();
-
-        assert!(rx.try_recv().is_ok(), "catch-up must run the LSP phase");
-        assert!(
-            latch.load(std::sync::atomic::Ordering::Acquire),
-            "serverless pass must latch lsp_unavailable"
-        );
-        assert!(
-            db.lock().unwrap().has_heuristic_exhausted().unwrap(),
-            "backlog must be re-sealed after a serverless pass"
-        );
-        assert_eq!(out.edges_lsp_resolved, 0);
-    }
-
-    #[cfg(feature = "lsp")]
-    #[test]
-    fn catch_up_skips_when_latched() {
-        let (_tmp, root, db) = sealed_py_fixture();
-        let latch = std::sync::atomic::AtomicBool::new(true);
-        let mgr = Arc::new(Mutex::new(no_server_manager(&root)));
-        let (tx, mut rx) = tokio::sync::mpsc::channel(64);
-
-        let noop = indexer::IndexResult::default();
-        catch_up_lsp(&db, &mgr, &latch, &root, Some(tx), None, noop).unwrap();
-
-        assert!(
-            rx.try_recv().is_err(),
-            "latched server must not retry the catch-up"
-        );
-    }
-
-    #[cfg(feature = "lsp")]
-    #[test]
-    fn catch_up_skips_dirty_runs() {
-        // Dirty runs already did a global warm pass in index_with_optional_lsp.
-        let (_tmp, root, db) = sealed_py_fixture();
-        let latch = std::sync::atomic::AtomicBool::new(false);
-        let mgr = Arc::new(Mutex::new(no_server_manager(&root)));
-        let (tx, mut rx) = tokio::sync::mpsc::channel(64);
-
-        let dirty = indexer::IndexResult {
-            dirty_files: 1,
-            ..Default::default()
-        };
-        catch_up_lsp(&db, &mgr, &latch, &root, Some(tx), None, dirty).unwrap();
-
-        assert!(rx.try_recv().is_err(), "dirty run must skip the catch-up");
-        assert!(!latch.load(std::sync::atomic::Ordering::Acquire));
-    }
-
-    #[cfg(feature = "lsp")]
-    #[test]
-    fn catch_up_noop_without_sealed_backlog() {
-        let tmp = tempfile::tempdir().unwrap();
-        let root = tmp.path().canonicalize().unwrap().join("project");
-        std::fs::create_dir_all(&root).unwrap();
-        let db = Arc::new(Mutex::new(cartog_db::Database::open_memory().unwrap()));
-        let latch = std::sync::atomic::AtomicBool::new(false);
-        let mgr = Arc::new(Mutex::new(no_server_manager(&root)));
-        let (tx, mut rx) = tokio::sync::mpsc::channel(64);
-
-        let noop = indexer::IndexResult::default();
-        catch_up_lsp(&db, &mgr, &latch, &root, Some(tx), None, noop).unwrap();
-
-        assert!(rx.try_recv().is_err(), "nothing sealed → no LSP phase");
-        assert!(
-            !latch.load(std::sync::atomic::Ordering::Acquire),
-            "latch must not be set when the pass never ran"
-        );
-    }
-
-    #[test]
-    fn pid_file_acquire_failure_propagates() {
-        // Pointing pid_lock_dir at a regular file makes ProcessLock::acquire
-        // fail at create_dir_all; the error must surface to the caller so
-        // `cartog serve` aborts rather than silently running unlocked.
-        let blocker = tempfile::NamedTempFile::new().unwrap();
-        let opts = ServerOptions {
-            pid_lock_dir: Some(blocker.path().to_path_buf()),
-            pid_lock_slot: Some(SERVE_LOCK_SLOT.to_string()),
-        };
-        let err = acquire_serve_lock(&opts).unwrap_err();
-        assert!(
-            err.to_string().contains("serve PID lock"),
-            "error should mention the lock context, got: {err}"
-        );
-    }
-
-    // ── Promoter regression tests (review fix M-promoter) ──
-
-    fn promoter_args_for_test(
-        db: Arc<Mutex<Database>>,
-        role: Arc<AtomicRole>,
-        db_path: std::path::PathBuf,
-        state_dir: std::path::PathBuf,
-        primary: cartog_process_lock::ActiveLock,
-        pinned: Option<PinnedAttach>,
-    ) -> PromoterArgs {
-        promoter_args_with_slot(
-            db,
-            role,
-            db_path,
-            state_dir,
-            primary,
-            pinned,
-            SERVE_LOCK_SLOT,
-        )
-    }
-
-    fn promoter_args_with_slot(
-        db: Arc<Mutex<Database>>,
-        role: Arc<AtomicRole>,
-        db_path: std::path::PathBuf,
-        state_dir: std::path::PathBuf,
-        primary: cartog_process_lock::ActiveLock,
-        pinned: Option<PinnedAttach>,
-        serve_slot: &str,
-    ) -> PromoterArgs {
-        // Test embedding provider: the reconcile step on promotion needs
-        // SOMETHING to fingerprint. The promoter-test DBs are opened fresh and
-        // never reconciled, so reconcile takes the harmless backfill branch
-        // (stamps mock provider/model) rather than wiping a populated index.
-        // Mock avoids loading the real ONNX model.
-        PromoterArgs {
-            db,
-            role,
-            lock_cell: Arc::new(Mutex::new(None)),
-            watch_cell: Arc::new(Mutex::new(None)),
-            stale_cell: Arc::new(Mutex::new(None)),
-            watcher_active: Arc::new(std::sync::atomic::AtomicBool::new(false)),
-            embedding_provider: Arc::new(Mutex::new(test_provider())),
-            db_path: db_path.clone(),
-            state_dir,
-            serve_slot: serve_slot.to_string(),
-            watch_slot: serve_to_watch_slot(serve_slot).expect("test slot must be valid"),
-            cwd: std::env::current_dir().unwrap(),
-            primary,
-            pinned,
-            watch_requested: false,
-            rag_override: Some(false),
-            rag_config: rag::EmbeddingProviderConfig::default(),
-            redact: indexer::RedactionConfig::disabled(),
-            walk_filter: indexer::WalkFilter::unrestricted(),
-            // Very short for tests so the loop responds quickly.
-            poll_interval: std::time::Duration::from_millis(20),
-        }
-    }
-
-    #[tokio::test]
-    async fn promoter_abort_cancels_the_polling_task() {
-        // Regression for review fix M-promoter (d): dropping the JoinHandle
-        // does NOT cancel a tokio task — only abort() does. Without the
-        // abort in run_server's shutdown path, the promoter could keep
-        // polling for up to one poll_interval after run_server returns
-        // and even promote during that window. We assert that abort()
-        // really terminates the task within a small bounded time.
-        let _serial = test_validate_call_counter::SERIAL.lock().await;
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let state_dir = dir.path().join("state");
-        // Materialize a DB so open_readonly can attach.
-        {
-            let _ = Database::open(&db_path, 384).unwrap();
-        }
-        let db = Arc::new(Mutex::new(Database::open_readonly(&db_path).unwrap()));
-        let role = Arc::new(AtomicRole::new(Role::ReadOnly));
-        let pinned = db.lock().unwrap().pinned_attach().cloned();
-        // Pretend the primary is our own process (so liveness reports
-        // true and the promoter just keeps polling forever, never
-        // promoting). This isolates the test to the abort behavior.
-        let primary = cartog_process_lock::ActiveLock {
-            slot: SERVE_LOCK_SLOT.to_string(),
-            pid: std::process::id(),
-            start_time: cartog_process_lock::process_start_time(std::process::id()),
-        };
-        let args = promoter_args_for_test(db, role, db_path, state_dir, primary, pinned);
-        let handle = tokio::task::spawn(promoter_task(args));
-
-        // Let it poll a few times.
-        tokio::time::sleep(std::time::Duration::from_millis(80)).await;
-        assert!(!handle.is_finished(), "promoter must keep polling");
-        handle.abort();
-        // Allow a brief moment for the runtime to cancel.
-        tokio::time::sleep(std::time::Duration::from_millis(40)).await;
-        assert!(
-            handle.is_finished(),
-            "abort must terminate the promoter task"
-        );
-    }
-
-    #[test]
-    fn validate_pinned_state_returns_ok_when_pin_is_none() {
-        let _serial = test_validate_call_counter::SERIAL.blocking_lock();
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        {
-            let _ = Database::open(&db_path, 384).unwrap();
-        }
-        validate_pinned_state(&db_path, None).expect("no pin must validate trivially");
-    }
-
-    #[test]
-    fn validate_pinned_state_detects_none_vs_some_drift() {
-        // A brand-new DB never reconciled has no provider/model in
-        // metadata, so a read-only attach captures `pinned.embedding =
-        // None`. If the primary subsequently runs `rag index` and stamps
-        // provider/model, the secondary must detect this as drift — pin
-        // was None, disk is now Some(...). This path is distinct from
-        // `validate_pinned_state_detects_drift`, which seeds provider+
-        // model BEFORE the attach so the pin is Some(...) and only the
-        // Some-vs-Some inequality is exercised.
-        let _serial = test_validate_call_counter::SERIAL.blocking_lock();
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let pinned = {
-            let _ = Database::open(&db_path, 384).unwrap();
-            Database::open_readonly(&db_path)
-                .unwrap()
-                .pinned_attach()
-                .cloned()
-        };
-        assert!(
-            pinned.as_ref().is_some_and(|p| p.embedding.is_none()),
-            "pin against an un-reconciled DB must capture embedding = None",
-        );
-        // Primary stamps provider+model.
-        {
-            let mutator = Database::open(&db_path, 384).unwrap();
-            mutator.set_metadata("embedding_provider", "local").unwrap();
-            mutator
-                .set_metadata("embedding_model", "BGE-small-en-v1.5")
-                .unwrap();
-        }
-        let err = validate_pinned_state(&db_path, pinned.as_ref())
-            .expect_err("None pin vs Some on disk must surface as drift");
-        assert!(
-            err.to_string().contains("DB metadata changed"),
-            "drift error should name the metadata change, got: {err}"
-        );
-    }
-
-    #[test]
-    fn validate_pinned_state_detects_drift() {
-        // Drift is exercised via the embedding fingerprint (not
-        // schema_version) because `open_readonly` rejects schema_version
-        // mismatch *before* this helper's comparison runs.
-        //
-        // We seed provider+model BEFORE the read-only attach so the pin
-        // captures `embedding = Some(local, BGE-..., 384)`, then mutate
-        // to a different `Some(ollama, nomic, 384)`. This exercises the
-        // Some vs Some inequality directly — not the None vs Some accident
-        // that would silently stop firing if `Database::open` ever seeded
-        // default provider/model values.
-        let _serial = test_validate_call_counter::SERIAL.blocking_lock();
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let pinned = {
-            let db = Database::open(&db_path, 384).unwrap();
-            db.set_metadata("embedding_provider", "local").unwrap();
-            db.set_metadata("embedding_model", "BGE-small-en-v1.5")
-                .unwrap();
-            drop(db);
-            Database::open_readonly(&db_path)
-                .unwrap()
-                .pinned_attach()
-                .cloned()
-        };
-        assert!(
-            pinned.as_ref().and_then(|p| p.embedding.as_ref()).is_some(),
-            "pin must capture Some(...) so the test exercises Some vs Some drift",
-        );
-        // Another writer rewrites provider + model under us.
-        {
-            let mutator = Database::open(&db_path, 384).unwrap();
-            mutator
-                .set_metadata("embedding_provider", "ollama")
-                .unwrap();
-            mutator
-                .set_metadata("embedding_model", "nomic-embed-text")
-                .unwrap();
-        }
-        let err = validate_pinned_state(&db_path, pinned.as_ref())
-            .expect_err("divergent disk state must surface as Err");
-        let msg = err.to_string();
-        assert!(
-            msg.contains("DB metadata changed"),
-            "error message should name the drift, got: {msg}"
-        );
-    }
-
-    #[tokio::test]
-    async fn promoter_acquires_db_scoped_slot_from_args() {
-        // Regression: promoter_task must acquire the slot carried in
-        // `args.serve_slot`, not a hardcoded `SERVE_LOCK_SLOT`. The
-        // previous test scaffolding hardcoded the global slot, so a
-        // refactor that reverts the acquire call to `SERVE_LOCK_SLOT`
-        // would have shipped green. This test passes a non-global serve
-        // slot and asserts the on-disk PID filename matches.
-        let _serial = test_validate_call_counter::SERIAL.lock().await;
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let state_dir = dir.path().join("state");
-        std::fs::create_dir_all(&state_dir).unwrap();
-        {
-            let _ = Database::open(&db_path, 384).unwrap();
-        }
-        let reader_db = Database::open_readonly(&db_path).unwrap();
-        let pinned = reader_db.pinned_attach().cloned();
-        let db = Arc::new(Mutex::new(reader_db));
-        let role = Arc::new(AtomicRole::new(Role::ReadOnly));
-        // Dead primary holding the scoped slot we want the promoter to
-        // claim. Slot name must match exactly what the promoter acquires
-        // so the original `serve-fa11`.pid file (planted below) gets
-        // reclaimed.
-        let scoped_slot = "serve-fa11ed7e57c0fed5";
-        let primary_pid_path = state_dir.join(format!("{scoped_slot}.pid"));
-        std::fs::write(&primary_pid_path, "4194304\n0\n").unwrap();
-        let primary = cartog_process_lock::ActiveLock {
-            slot: scoped_slot.to_string(),
-            pid: 4_194_304,
-            start_time: None,
-        };
-
-        let args = promoter_args_with_slot(
-            Arc::clone(&db),
-            Arc::clone(&role),
-            db_path,
-            state_dir.clone(),
-            primary,
-            pinned,
-            scoped_slot,
-        );
-        // Keep the lock_cell Arc alive past the task so the acquired
-        // ProcessLock isn't dropped (and the PID file removed) before we
-        // assert on it. The production run_server lifecycle does the
-        // same: it holds lock_cell for the whole server lifetime.
-        let lock_cell = Arc::clone(&args.lock_cell);
-
-        let handle = tokio::task::spawn(promoter_task(args));
-        // Give the promoter time to notice primary-gone and acquire.
-        tokio::time::sleep(std::time::Duration::from_millis(200)).await;
-
-        assert_eq!(
-            role.load(),
-            Role::Primary,
-            "promoter must flip role after acquiring the scoped slot"
-        );
-        // The on-disk PID file must use the scoped slot, not the global.
-        assert!(
-            primary_pid_path.exists(),
-            "expected promoter to acquire {primary_pid_path:?}"
-        );
-        let global_path = state_dir.join(format!("{SERVE_LOCK_SLOT}.pid"));
-        assert!(
-            !global_path.exists(),
-            "promoter must NOT acquire the global slot ({global_path:?})"
-        );
-
-        // Drop the held lock explicitly so the temp dir teardown is clean.
-        {
-            let mut guard = lock_cell.lock().unwrap();
-            *guard = None;
-        }
-        handle.abort();
-        let _ = handle.await;
-    }
-
-    #[tokio::test]
-    async fn promoter_spawns_watcher_with_db_scoped_watch_slot() {
-        // Regression for review finding #2: promoter_task must claim the
-        // watch slot carried in `args.watch_slot`, not a hardcoded
-        // `WATCH_LOCK_SLOT`. The previous regression test only verified
-        // the serve-slot acquire path; a refactor that reverts the
-        // post-promotion watcher's `config.pid_lock_slot = Some(args
-        // .watch_slot.clone())` to the global constant would have shipped
-        // green. This test sets `watch_requested = true`, lets the
-        // promoter spawn a watcher, and asserts the watcher claims the
-        // scoped slot on disk.
-        let _serial = test_validate_call_counter::SERIAL.lock().await;
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let state_dir = dir.path().join("state");
-        std::fs::create_dir_all(&state_dir).unwrap();
-        {
-            let _ = Database::open(&db_path, 384).unwrap();
-        }
-        let reader_db = Database::open_readonly(&db_path).unwrap();
-        let pinned = reader_db.pinned_attach().cloned();
-        let db = Arc::new(Mutex::new(reader_db));
-        let role = Arc::new(AtomicRole::new(Role::ReadOnly));
-        let scoped_slot = "serve-fa11ed7e57c0fed5";
-        let primary_pid_path = state_dir.join(format!("{scoped_slot}.pid"));
-        std::fs::write(&primary_pid_path, "4194304\n0\n").unwrap();
-        let primary = cartog_process_lock::ActiveLock {
-            slot: scoped_slot.to_string(),
-            pid: 4_194_304,
-            start_time: None,
-        };
-
-        let mut args = promoter_args_with_slot(
-            Arc::clone(&db),
-            Arc::clone(&role),
-            db_path,
-            state_dir.clone(),
-            primary,
-            pinned,
-            scoped_slot,
-        );
-        // Enable the watcher-spawn path. RAG stays off and watch root
-        // (cwd, captured by the helper) is the cartog crate dir — fine
-        // for spawning the watcher; we tear it down immediately.
-        args.watch_requested = true;
-        let lock_cell = Arc::clone(&args.lock_cell);
-        let watch_cell = Arc::clone(&args.watch_cell);
-
-        let handle = tokio::task::spawn(promoter_task(args));
-
-        // The expected watcher PID file uses the SCOPED watch slot derived
-        // via `serve_to_watch_slot(scoped_slot)`.
-        let expected_watch_slot = serve_to_watch_slot(scoped_slot).expect("scoped slot is valid");
-        let expected_watch_pid = state_dir.join(format!("{expected_watch_slot}.pid"));
-        let global_watch_pid = state_dir.join(format!("{}.pid", watch::WATCH_LOCK_SLOT));
-
-        // Poll for up to 5s in 50ms increments — the watcher startup walks
-        // the cwd, sweeps stale locks, and creates the debouncer; on a
-        // loaded CI machine the cumulative cost can exceed any small fixed
-        // sleep, so a fixed sleep here was a flake source.
-        let deadline = std::time::Instant::now() + std::time::Duration::from_secs(5);
-        while !expected_watch_pid.exists() && std::time::Instant::now() < deadline {
-            tokio::time::sleep(std::time::Duration::from_millis(50)).await;
-        }
-        assert!(
-            expected_watch_pid.exists(),
-            "watcher should have claimed the scoped slot at {expected_watch_pid:?}"
-        );
-        assert!(
-            !global_watch_pid.exists(),
-            "watcher must NOT claim the global slot ({global_watch_pid:?})"
-        );
-
-        // Clean shutdown so the temp dir teardown succeeds.
-        {
-            let mut wguard = watch_cell.lock().unwrap();
-            if let Some(handle) = wguard.take() {
-                handle.stop();
-            }
-        }
-        {
-            let mut guard = lock_cell.lock().unwrap();
-            *guard = None;
-        }
-        handle.abort();
-        let _ = handle.await;
-    }
-
-    #[tokio::test]
-    async fn promoter_aborts_when_state_diverges_after_acquire() {
-        // Integration smoke. The post-acquire branch logic is covered
-        // by `validate_pinned_state_detects_drift`; this test verifies
-        // the promoter wires drift detection to a clean exit (role stays
-        // ReadOnly, task finishes).
-        let _serial = test_validate_call_counter::SERIAL.lock().await;
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let state_dir = dir.path().join("state");
-        std::fs::create_dir_all(&state_dir).unwrap();
-        {
-            let _ = Database::open(&db_path, 384).unwrap();
-        }
-        let reader_db = Database::open_readonly(&db_path).unwrap();
-        let pinned = reader_db.pinned_attach().cloned();
-        let db = Arc::new(Mutex::new(reader_db));
-        let role = Arc::new(AtomicRole::new(Role::ReadOnly));
-        // Pretend primary is dead (no such PID).
-        let primary = cartog_process_lock::ActiveLock {
-            slot: SERVE_LOCK_SLOT.to_string(),
-            pid: 4_194_304,
-            start_time: None,
-        };
-        // Mutate the DB metadata under the reader.
-        {
-            let mutator = Database::open(&db_path, 384).unwrap();
-            mutator.set_metadata("schema_version", "9999").unwrap();
-        }
-
-        let args = promoter_args_for_test(
-            Arc::clone(&db),
-            Arc::clone(&role),
-            db_path,
-            state_dir,
-            primary,
-            pinned,
-        );
-        let handle = tokio::task::spawn(promoter_task(args));
-        // Give the promoter one tick.
-        tokio::time::sleep(std::time::Duration::from_millis(80)).await;
-
-        // The promoter should have noticed primary-gone, validated, seen
-        // drift, and exited (return). Role must NOT be Primary.
-        assert!(handle.is_finished(), "promoter must exit on drift");
-        let _ = handle.await;
-        assert_eq!(
-            role.load(),
-            Role::ReadOnly,
-            "drifted DB must not flip role to Primary"
-        );
-    }
-
-    #[tokio::test]
-    async fn promoter_loops_on_transient_open_failure() {
-        // Regression for review fix M-promoter (b): pre-fix, an
-        // open_existing_rw failure caused the promoter to `return`,
-        // disabling promotion forever even if the next poll would
-        // succeed. The fix loops on transient failures.
-        //
-        // We can exercise the "open fails -> loop" path by deleting the
-        // DB file entirely between the validate and the open_existing_rw
-        // call. open_existing_rw will fail; the promoter should drop
-        // the lock and try again on the next tick (where it'll fail
-        // validation, since the DB is missing, and exit cleanly).
-        //
-        // The key contract is: we don't return on the first
-        // open_existing_rw failure — we drop the lock and loop. We
-        // assert that by checking the lock file does not persist after
-        // a failed promotion attempt.
-        let _serial = test_validate_call_counter::SERIAL.lock().await;
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let state_dir = dir.path().join("state");
-        std::fs::create_dir_all(&state_dir).unwrap();
-        {
-            let _ = Database::open(&db_path, 384).unwrap();
-        }
-        let reader = Database::open_readonly(&db_path).unwrap();
-        let pinned = reader.pinned_attach().cloned();
-        let db = Arc::new(Mutex::new(reader));
-        let role = Arc::new(AtomicRole::new(Role::ReadOnly));
-        let primary = cartog_process_lock::ActiveLock {
-            slot: SERVE_LOCK_SLOT.to_string(),
-            pid: 4_194_304,
-            start_time: None,
-        };
-
-        let args = promoter_args_for_test(
-            Arc::clone(&db),
-            Arc::clone(&role),
-            db_path.clone(),
-            state_dir.clone(),
-            primary,
-            pinned,
-        );
-        let handle = tokio::task::spawn(promoter_task(args));
-        // Give the loop a moment to enter its first tick, then yank the DB.
-        tokio::time::sleep(std::time::Duration::from_millis(15)).await;
-        std::fs::remove_file(&db_path).unwrap();
-        // The promoter should now either: (a) loop on validate-failure
-        // and never acquire, or (b) acquire then fail open and drop the
-        // lock + loop. Either way the role stays ReadOnly and the lock
-        // file is not left behind.
-        tokio::time::sleep(std::time::Duration::from_millis(150)).await;
-        handle.abort();
-        let _ = handle.await;
-        assert_eq!(
-            role.load(),
-            Role::ReadOnly,
-            "promoter must not flip role under transient failure"
-        );
-        let lock_path = state_dir.join("serve.pid");
-        assert!(
-            !lock_path.exists(),
-            "promoter must release the lock on failure (not strand serve.pid)"
-        );
-    }
-
-    #[tokio::test]
-    async fn promoter_runs_validate_pinned_state_both_before_and_after_acquire() {
-        // Regression for the M-promoter (c) fix at the test layer: the
-        // promoter must call `validate_pinned_state` BOTH before and
-        // after `ProcessLock::acquire`. We need an assertion that catches
-        // deleting EITHER call site:
-        //
-        // - `calls >= 2` alone is insufficient: pre-acquire validate
-        //   bumps on every tick, so over multiple ticks the count climbs
-        //   past 2 even if the post-acquire site is gone.
-        // - `role == Primary` alone is insufficient: the role swap
-        //   succeeds even if pre-acquire validate is skipped entirely.
-        //
-        // Asserting BOTH catches either deletion: post-acquire validate
-        // gates the role swap (so Primary => post-acquire ran on at
-        // least one tick), and we additionally require `calls >= 2` to
-        // prove a second validate fired in the promote path.
-        let _serial = test_validate_call_counter::SERIAL.lock().await;
-        test_validate_call_counter::reset();
-
-        let dir = tempfile::TempDir::new().unwrap();
-        let db_path = dir.path().join("test.db");
-        let state_dir = dir.path().join("state");
-        std::fs::create_dir_all(&state_dir).unwrap();
-        {
-            let _ = Database::open(&db_path, 384).unwrap();
-        }
-        let reader = Database::open_readonly(&db_path).unwrap();
-        let pinned = reader.pinned_attach().cloned();
-        let db = Arc::new(Mutex::new(reader));
-        let role = Arc::new(AtomicRole::new(Role::ReadOnly));
-        // Pretend the primary is dead so the promoter exits the
-        // primary-alive `continue` branch and reaches the validate calls.
-        let primary = cartog_process_lock::ActiveLock {
-            slot: SERVE_LOCK_SLOT.to_string(),
-            pid: 4_194_304,
-            start_time: None,
-        };
-
-        let args = promoter_args_for_test(
-            Arc::clone(&db),
-            Arc::clone(&role),
-            db_path,
-            state_dir,
-            primary,
-            pinned,
-        );
-        let handle = tokio::task::spawn(promoter_task(args));
-
-        // Bounded poll instead of a fixed wall-clock sleep: keeps the
-        // test fast on a healthy box and tolerates a stalled CI runner
-        // up to the 2s deadline. We wait for BOTH conditions before
-        // asserting — reading them once after a fixed sleep was the
-        // source of the previous CI-flake risk.
-        let deadline = std::time::Instant::now() + std::time::Duration::from_secs(2);
-        loop {
-            if role.load() == Role::Primary && test_validate_call_counter::snapshot() >= 2 {
-                break;
-            }
-            if std::time::Instant::now() >= deadline {
-                break;
-            }
-            tokio::time::sleep(std::time::Duration::from_millis(20)).await;
-        }
-        handle.abort();
-        let _ = handle.await;
-
-        let calls = test_validate_call_counter::snapshot();
-        assert_eq!(
-            role.load(),
-            Role::Primary,
-            "post-acquire validate gates the role swap; missing role flip implies the post-acquire site never ran"
-        );
-        assert!(
-            calls >= 2,
-            "validate_pinned_state must run twice per promotion tick (pre + post acquire), saw {calls}"
-        );
-    }
-
-    // ── Read-tool handler tests over a real indexed DB ──
-    //
-    // Build a CartogServer over a temp DB pre-populated by indexing a small
-    // Python fixture, then drive the async read handlers directly. This
-    // exercises the real MCP dispatch (param parsing, error mapping, the
-    // tool_response / tool_response_named integration) over real query
-    // results, not mocks.
-
-    const FIXTURE_SRC: &str = "\
-class Animal:
-    def speak(self):
-        return helper()
-
-
-class Dog(Animal):
-    def speak(self):
-        return helper()
-
-
-def helper():
-    return 42
-
-
-def main():
-    d = Dog()
-    return d.speak()
-";
-
-    /// Index `FIXTURE_SRC` as `lib.py` into a temp dir, then return a primary
-    /// server opened over the resulting DB. The TempDir is returned so the
-    /// caller keeps it alive for the test's duration.
-    fn indexed_server() -> (tempfile::TempDir, CartogServer) {
-        let tmp = tempfile::TempDir::new().unwrap();
-        // The index root must not be a dot-prefixed dir: the walker prunes any
-        // entry whose name starts with '.', and TempDir names start with ".tmp".
-        let root = tmp.path().join("project");
-        std::fs::create_dir(&root).unwrap();
-        std::fs::write(root.join("lib.py"), FIXTURE_SRC).unwrap();
-        let db_path = tmp.path().join("cartog.db");
-        let provider = test_provider();
-        {
-            let db = Database::open(&db_path, provider.dimension()).unwrap();
-            indexer::index_directory(
-                &db,
-                &root,
-                true,
-                false,
-                None,
-                None,
-                indexer::RedactionConfig::disabled(),
-                &std::collections::HashMap::new(),
-                &indexer::WalkFilter::unrestricted(),
-            )
-            .expect("fixture indexes");
-        }
-        let server = CartogServer::new_with_provider(
-            &db_path,
-            provider,
-            indexer::RedactionConfig::disabled(),
-            indexer::WalkFilter::unrestricted(),
-            Role::Primary,
-        )
-        .expect("server constructs");
-        (tmp, server)
-    }
-
-    /// Extract the text payload of a successful single-content tool result.
-    fn result_text(result: &CallToolResult) -> String {
-        result
-            .content
-            .first()
-            .and_then(|c| c.as_text())
-            .map(|t| t.text.clone())
-            .expect("tool result has text content")
-    }
-
-    #[tokio::test]
-    async fn outline_lists_symbols_in_file() {
-        let (_dir, server) = indexed_server();
-        let result = server
-            .cartog_outline(Parameters(OutlineParams {
-                file: "lib.py".to_string(),
-            }))
-            .await
-            .expect("outline succeeds");
-        let text = result_text(&result);
-        assert!(
-            text.contains("Animal"),
-            "outline should list the Animal class"
-        );
-        assert!(
-            text.contains("helper"),
-            "outline should list the helper function"
-        );
-    }
-
-    #[tokio::test]
-    async fn outline_unknown_file_returns_empty_array() {
-        let (_dir, server) = indexed_server();
-        let result = server
-            .cartog_outline(Parameters(OutlineParams {
-                file: "nonexistent.py".to_string(),
-            }))
-            .await
-            .expect("outline of unknown file is not an error");
-        let text = result_text(&result);
-        assert!(
-            text.trim_start().starts_with('['),
-            "empty outline is a JSON array"
-        );
-    }
-
-    #[tokio::test]
-    async fn refs_finds_callers_of_helper() {
-        let (_dir, server) = indexed_server();
-        let result = server
-            .cartog_refs(Parameters(RefsParams {
-                name: "helper".to_string(),
-                kind: None,
-            }))
-            .await
-            .expect("refs succeeds");
-        let text = result_text(&result);
-        assert!(text.contains("helper"), "refs to helper should mention it");
-    }
-
-    #[tokio::test]
-    async fn refs_rejects_invalid_edge_kind() {
-        let (_dir, server) = indexed_server();
-        let err = server
-            .cartog_refs(Parameters(RefsParams {
-                name: "helper".to_string(),
-                kind: Some("bogus".to_string()),
-            }))
-            .await
-            .expect_err("invalid edge kind must be rejected");
-        assert!(
-            err.message.contains("invalid edge kind"),
-            "error should name the invalid kind, got: {}",
-            err.message
-        );
-    }
-
-    #[tokio::test]
-    async fn refs_unknown_name_suggests_near_matches() {
-        let (_dir, server) = indexed_server();
-        // "helpe" is one char off "helper" — should trigger did-you-mean.
-        let result = server
-            .cartog_refs(Parameters(RefsParams {
-                name: "helpe".to_string(),
-                kind: None,
-            }))
-            .await
-            .expect("refs of near-miss name succeeds");
-        let text = result_text(&result);
-        assert!(
-            text.contains("Did you mean") && text.contains("helper"),
-            "near-miss should suggest helper, got: {text}"
-        );
-    }
-
-    #[tokio::test]
-    async fn callees_traces_calls_from_main() {
-        let (_dir, server) = indexed_server();
-        let result = server
-            .cartog_callees(Parameters(CalleesParams {
-                name: "main".to_string(),
-            }))
-            .await
-            .expect("callees succeeds");
-        let text = result_text(&result);
-        assert!(
-            text.trim_start().starts_with('['),
-            "callees returns a JSON array"
-        );
-    }
-
-    #[tokio::test]
-    async fn impact_clamps_depth_and_returns_array() {
-        let (_dir, server) = indexed_server();
-        let result = server
-            .cartog_impact(Parameters(ImpactParams {
-                name: "helper".to_string(),
-                depth: Some(999), // clamped to MAX_IMPACT_DEPTH internally
-            }))
-            .await
-            .expect("impact succeeds");
-        let text = result_text(&result);
-        assert!(
-            text.trim_start().starts_with('['),
-            "impact returns a JSON array"
-        );
-    }
-
-    #[tokio::test]
-    async fn trace_finds_path_from_speak_to_helper() {
-        let (_dir, server) = indexed_server();
-        let result = server
-            .cartog_trace(Parameters(TraceParams {
-                from: "speak".to_string(),
-                to: "helper".to_string(),
-                depth: Some(8),
-            }))
-            .await
-            .expect("trace succeeds");
-        let text = result_text(&result);
-        assert!(text.contains("\"found\": true"), "path exists: {text}");
-        assert!(text.contains("helper"), "hop should reach helper: {text}");
-    }
-
-    #[tokio::test]
-    async fn trace_reports_no_path_when_unreachable() {
-        let (_dir, server) = indexed_server();
-        let result = server
-            .cartog_trace(Parameters(TraceParams {
-                from: "helper".to_string(),
-                to: "speak".to_string(),
-                depth: Some(8),
-            }))
-            .await
-            .expect("trace succeeds");
-        let text = result_text(&result);
-        assert!(text.contains("\"found\": false"), "no path: {text}");
-    }
-
-    #[tokio::test]
-    async fn trace_hop_includes_body_when_content_indexed() {
-        let (_dir, server) = indexed_server();
-        // Seed RAG content for every `speak` symbol (the hop sources on the
-        // speak→helper path), so the hop body is populated.
-        {
-            let db = server.db.lock().unwrap();
-            for sym in db.search("speak", None, None, 10).unwrap() {
-                db.upsert_symbol_content(
-                    &sym.id,
-                    "speak",
-                    "def speak(self):\n    return helper()",
-                    "// method speak",
-                )
-                .unwrap();
-            }
-        }
-        let result = server
-            .cartog_trace(Parameters(TraceParams {
-                from: "speak".to_string(),
-                to: "helper".to_string(),
-                depth: Some(8),
-            }))
-            .await
-            .expect("trace succeeds");
-        let text = result_text(&result);
-        assert!(
-            text.contains("\"body\""),
-            "hop carries an inline body when content is indexed: {text}"
-        );
-    }
-
-    #[tokio::test]
-    async fn context_bundles_relevant_symbols_for_a_task() {
-        let (_dir, server) = indexed_server();
-        let result = server
-            .cartog_context(Parameters(ContextParams {
-                task: "speak".to_string(),
-                tokens: Some(6000),
-            }))
-            .await
-            .expect("context succeeds");
-        let text = result_text(&result);
-        assert!(text.contains("\"task\": \"speak\""), "echoes task: {text}");
-        assert!(text.contains("\"entries\""), "returns entries: {text}");
-    }
-
-    #[tokio::test]
-    async fn rag_search_prepends_banner_when_embeddings_pending() {
-        let (_dir, server) = indexed_server();
-        // Simulate a live watcher with pending embeddings.
-        let stale = cartog_watch::StaleState::new();
-        stale.note_reindex(0, 7);
-        *server.stale.lock().unwrap() = Some(stale);
-        server
-            .watcher_active
-            .store(true, std::sync::atomic::Ordering::Relaxed);
-
-        let result = server
-            .cartog_rag_search(Parameters(RagSearchParams {
-                query: "helper".to_string(),
-                kind: None,
-                limit: Some(5),
-            }))
-            .await
-            .expect("rag search succeeds");
-        let text = result_text(&result);
-        assert!(
-            text.starts_with("⚠️") && text.contains("re-embedding"),
-            "banner prepended: {text}"
-        );
-    }
-
-    #[tokio::test]
-    async fn no_banner_without_active_watcher() {
-        let (_dir, server) = indexed_server();
-        // Stale state present but watcher_active is false (degraded/no watcher).
-        let stale = cartog_watch::StaleState::new();
-        stale.note_reindex(0, 7);
-        *server.stale.lock().unwrap() = Some(stale);
-
-        let result = server
-            .cartog_rag_search(Parameters(RagSearchParams {
-                query: "helper".to_string(),
-                kind: None,
-                limit: Some(5),
-            }))
-            .await
-            .expect("rag search succeeds");
-        assert!(!result_text(&result).starts_with("⚠️"), "no banner");
-    }
-
-    #[tokio::test]
-    async fn context_rejects_empty_task() {
-        let (_dir, server) = indexed_server();
-        let err = server
-            .cartog_context(Parameters(ContextParams {
-                task: String::new(),
-                tokens: None,
-            }))
-            .await
-            .expect_err("empty task must be rejected");
-        assert!(
-            err.message.contains("cannot be empty"),
-            "got: {}",
-            err.message
-        );
-    }
-
-    #[tokio::test]
-    async fn hierarchy_reports_dog_extends_animal() {
-        let (_dir, server) = indexed_server();
-        let result = server
-            .cartog_hierarchy(Parameters(HierarchyParams {
-                name: "Dog".to_string(),
-            }))
-            .await
-            .expect("hierarchy succeeds");
-        let text = result_text(&result);
-        assert!(
-            text.contains("Animal"),
-            "Dog's hierarchy should reach Animal: {text}"
-        );
-    }
-
-    #[tokio::test]
-    async fn deps_lists_file_imports() {
-        let (_dir, server) = indexed_server();
-        let result = server
-            .cartog_deps(Parameters(DepsParams {
-                file: "lib.py".to_string(),
-            }))
-            .await
-            .expect("deps succeeds");
-        let text = result_text(&result);
-        assert!(
-            text.trim_start().starts_with('['),
-            "deps returns a JSON array"
-        );
-    }
-
-    #[tokio::test]
-    async fn search_finds_symbol_by_partial_name() {
-        let (_dir, server) = indexed_server();
-        let result = server
-            .cartog_search(Parameters(SearchParams {
-                query: "Anim".to_string(),
-                kind: None,
-                file: None,
-                limit: None,
-            }))
-            .await
-            .expect("search succeeds");
-        let text = result_text(&result);
-        assert!(
-            text.contains("Animal"),
-            "search for 'Anim' should find Animal"
-        );
-    }
-
-    #[tokio::test]
-    async fn search_rejects_empty_query() {
-        let (_dir, server) = indexed_server();
-        let err = server
-            .cartog_search(Parameters(SearchParams {
-                query: String::new(),
-                kind: None,
-                file: None,
-                limit: None,
-            }))
-            .await
-            .expect_err("empty query must be rejected");
-        assert!(
-            err.message.contains("query cannot be empty"),
-            "got: {}",
-            err.message
-        );
-    }
-
-    #[tokio::test]
-    async fn search_rejects_invalid_kind() {
-        let (_dir, server) = indexed_server();
-        let err = server
-            .cartog_search(Parameters(SearchParams {
-                query: "Animal".to_string(),
-                kind: Some("nonsense".to_string()),
-                file: None,
-                limit: None,
-            }))
-            .await
-            .expect_err("invalid symbol kind must be rejected");
-        assert!(
-            err.message.contains("invalid symbol kind"),
-            "got: {}",
-            err.message
-        );
-    }
-
-    #[tokio::test]
-    async fn stats_reports_role_and_symbol_count() {
-        let (_dir, server) = indexed_server();
-        let result = server.cartog_stats().await.expect("stats succeeds");
-        let text = result_text(&result);
-        let value: serde_json::Value = serde_json::from_str(&text).expect("stats is JSON");
-        assert_eq!(
-            value["role"], "primary",
-            "primary server reports primary role"
-        );
-        assert!(
-            value["num_symbols"].as_u64().unwrap_or(0) > 0,
-            "indexed fixture has symbols: {text}"
-        );
-    }
-
-    #[tokio::test]
-    async fn map_returns_files_and_top_symbols() {
-        let (_dir, server) = indexed_server();
-        let result = server
-            .cartog_map(Parameters(MapParams { limit: Some(10) }))
-            .await
-            .expect("map succeeds");
-        let text = result_text(&result);
-        assert!(
-            text.contains("lib.py"),
-            "map should list the indexed file: {text}"
-        );
-    }
-
-    #[tokio::test]
-    async fn read_tools_count_toward_query_log() {
-        let (_dir, server) = indexed_server();
-        let _ = server
-            .cartog_search(Parameters(SearchParams {
-                query: "helper".to_string(),
-                kind: None,
-                file: None,
-                limit: None,
-            }))
-            .await
-            .expect("search succeeds");
-        let result = server.cartog_stats().await.expect("stats succeeds");
-        let value: serde_json::Value =
-            serde_json::from_str(&result_text(&result)).expect("stats is JSON");
-        // stats itself plus the prior search both log; the field exists once
-        // any read tool has run against a populated index.
-        assert!(value.get("num_symbols").is_some(), "stats shape is intact");
-    }
-}
+mod tests;
